Rename component traits to extensions
