+++ src/main/java/nerdhub/cardinal/components/api/event/ItemComponentCallback.java
@@ -23,7 +23,7 @@
 package nerdhub.cardinal.components.api.event;
 
 import nerdhub.cardinal.components.api.component.ComponentContainer;
-import nerdhub.cardinal.components.api.component.trait.CloneableComponent;
+import nerdhub.cardinal.components.api.component.extension.CloneableComponent;
 import nerdhub.cardinal.components.api.util.component.ItemComponent;
 import nerdhub.cardinal.components.internal.CardinalEventsInternals;
 import nerdhub.cardinal.components.internal.ItemCaller;
